examples/python-inference/requirements.txt
httpx==0.27.0
rerun-sdk
kornia_rs